sdk/code/python/requirements.txt
grpc-client-utils==1.0.0
grpcio-tools>=1.67
protobuf>=5.28.3